internal: transport, Move protocol-specific test utils to each package
