--- internal/transport/ssh/test/test_utils.go
@@ -1,83 +1,0 @@
-package test
-
-import (
-	"fmt"
-	"io"
-	"os/exec"
-	"runtime"
-	"strings"
-	"sync"
-
-	"github.com/gliderlabs/ssh"
-)
-
-func HandlerSSH(s ssh.Session) {
-	cmd, stdin, stderr, stdout, err := buildCommand(s.Command())
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	if err := cmd.Start(); err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	go func() {
-		defer stdin.Close()
-		io.Copy(stdin, s)
-	}()
-
-	var wg sync.WaitGroup
-	wg.Add(2)
-
-	go func() {
-		defer wg.Done()
-		io.Copy(s.Stderr(), stderr)
-	}()
-
-	go func() {
-		defer wg.Done()
-		io.Copy(s, stdout)
-	}()
-
-	wg.Wait()
-
-	if err := cmd.Wait(); err != nil {
-		return
-	}
-
-}
-
-func buildCommand(c []string) (cmd *exec.Cmd, stdin io.WriteCloser, stderr, stdout io.ReadCloser, err error) {
-	if len(c) != 2 {
-		err = fmt.Errorf("invalid command")
-		return
-	}
-
-	// fix for Windows environments
-	var path string
-	if runtime.GOOS == "windows" {
-		path = strings.Replace(c[1], "/C:", "C:", 1)
-	} else {
-		path = c[1]
-	}
-
-	cmd = exec.Command(c[0], path)
-	stdout, err = cmd.StdoutPipe()
-	if err != nil {
-		return
-	}
-
-	stdin, err = cmd.StdinPipe()
-	if err != nil {
-		return
-	}
-
-	stderr, err = cmd.StderrPipe()
-	if err != nil {
-		return
-	}
-
-	return
-}
