+++ plumbing/transport/http/testdata/certs/server.crt
